Fix printing of invalid struct ifreq.ifr_flags
* sock.c (print_ifreq): Explicitly cast struct ifreq.ifr_flags
to unsigned short to avoid potential sign-extension bug when printing
invalid struct ifreq.ifr_flags.

+++ sock.c
@@ -97,7 +97,7 @@ print_ifreq(struct tcb *tcp, const unsigned int code, const long arg,
 	case SIOCSIFFLAGS:
 	case SIOCGIFFLAGS:
 		tprints("ifr_flags=");
-		printflags(iffflags, ifr->ifr_flags, "IFF_???");
+		printflags(iffflags, (unsigned short) ifr->ifr_flags, "IFF_???");
 		break;
 	case SIOCSIFMETRIC:
 	case SIOCGIFMETRIC:
